make num_of_rings uint8_t as in the rest of the code
this patch should have been part of 16fc7fffd8b6fd49a0cf2235971a914ab15a3d68

Reviewed-on: http://git.am.freescale.net:8181/38782

+++ host_driver/fsl_c2x0_crypto_layer.h
@@ -521,7 +521,7 @@ typedef struct fsl_crypto_dev {
 	uint8_t max_pri_level;
 	fsl_priority_queue_t pri_queue[MAX_PRIORITY_LEVELS];
 
-	uint32_t num_of_rings;
+	uint8_t num_of_rings;
 	fsl_h_rsrc_ring_pair_t *ring_pairs;
 
 	/* Holds the count of number of crypto dev sessions */
